completed-api

+++ src/services/cartItem.service.js
@@ -4,30 +4,61 @@ const User = require("../models/user.model");
 
 async function updateCartItem(userId, cartItemId, cartItemData) {
   try {
-    const item = await CartItem.findById(cartItemId);
-    const user = await userService.getUserById(item.userId);
+    const item = await CartItem.findById(cartItemId).populate("product");
+    
     if (!item) {
       return { status: 404, message: "Item not found" };
     }
+    
+    const user = await userService.findUserById(item.userId);
+    
     if (!user) {
       return { status: 404, message: "User not found" };
     }
+    
     if (user._id.toString() === userId.toString()) {
-      item.quantity = cartItemData.quantity;
-      item.price = item.quantity * item.product.price;
-      item.discountedPrice = item.quantity * item.product.discountedPrice;
+      const quantity = Number(cartItemData.quantity);
+      if (isNaN(quantity)) {
+        throw new Error('Invalid data for quantity');
+      }
+
+      let totalDiscountedPrice = 0;
+      let totalPrice = 0;
+
+      // Loop through each product to calculate total price and discounted price
+      item.product.forEach(prod => {
+        const pricePerUnit = Number(prod.price);
+        const discountedPricePerUnit = Number(prod.discountedPrice);
+
+        if (isNaN(pricePerUnit) || isNaN(discountedPricePerUnit)) {
+          throw new Error('Invalid data for price or discountedPrice');
+        }
+
+        totalPrice += quantity * pricePerUnit;
+        totalDiscountedPrice += quantity * discountedPricePerUnit;
+      });
+
+      item.quantity = quantity;
+      item.price = totalPrice;
+      item.discountedPrice = totalDiscountedPrice;
+
       const updatedCartItem = await item.save();
-      return updatedCartItem();
+      return updatedCartItem;
     }
+    
     return { status: 401, message: "Unauthorized" };
   } catch (error) {
     throw new Error(error.message);
   }
 }
 
+
 async function removeCartItem(userId, cartItemId) {
   try {
     const cartItem = await CartItem.findById(cartItemId);
+    if(!cartItem){
+      return { status: 404, message: "Item not found" };
+    }
     if (cartItem.userId.toString() === userId.toString()) {
       await CartItem.findByIdAndDelete(cartItemId);
       return { status: 200, message: "Item removed successfully" };
